contrib: fix rpm spec file post install and uninstall actions
Earlier rpm actions where broken and wrong.  This commit fixes most of
the issues.  Additionally the installation will result to enabled state
by default if the state was not known after installation.  This is
assumed ease making systems being part of a load balanced group
automatically.  Unfortunately the automatic enabling can be harmful in
some environments so the bit where 'unknown' is enforced to 'enabled'
may need to be before building the rpm for such environments.

This commit also fixes various small issues in the spec file, such as
misleading version number in instruction comment very top of the file.

Signed-off-by: Sami Kerola <kerolasa@iki.fi>