add --no-scripts option
Especially the pre script failures can be quite annoying since the
state will not change until the script has ran manually.  There are
cases when operator knows what he is doing and the correct action is to
ignore the scripts.

This could happen for instance when a status is set to --disable and
iptable rules are set.  Then a server is rebooted, and the iptable
rules disappear but state is still disabled.  The --enable will try to
remove none existing iptable rules and fail.  Hence state change
without either removing the pre script or running the command with
--no-scripts is impossible.

Signed-off-by: Sami Kerola <kerolasa@iki.fi>