stop using SIGWINCH after introduction of --foreground
The --foreground allows running f5gs in a shell that is a window, and if
one changes size of a such window the signal is sent possibly without
users' awareness.  To avoid bug reports telling why state changed
seemingly without reason it is best to avoid the SIGWINCH all together.

Also make the association of signal and action to use symbolic
references, SIG_ENABLE etc, to make it as difficult as possible to mess
up code when changing a signal to another.

Signed-off-by: Sami Kerola <kerolasa@iki.fi>