restore portability, signalfd() might not be available
For systems, basically all but linux, that does not have signalfd() has
to fall back to sigwait().  It is probable this is not a big issue as the
section where the extra information provided by signalfd() depends on
systemd.  The systems without these facilities will simply print less
information what happen.

Signed-off-by: Sami Kerola <kerolasa@iki.fi>