add client side connection timeout
When trying to connect to destination that is blocked by firewall it
takes ages to timeout.  Since the f5gs is expected to answer very
quickly, when it works, reasonable timeout is one second.  This way
scanning a list of hosts to request their will be relatively fast, even
when there are unreachable targets.

Signed-off-by: Sami Kerola <kerolasa@iki.fi>