exit state_change_thread() thread gracefully
Let the thread exit, and join it at stop_server().  This way the msgrcv()
cannot try to get message from ipc queue that is removed.

Note that earlier use of pthread_cancel() was ineffective, because there
was no clean up handler in state_change_thread(), nor the msgrcv() block
was never released.

Signed-off-by: Sami Kerola <kerolasa@iki.fi>