check handle_request() thread can be created
Reasons why this might happen is system wide memory or available thread
limit full.  It is probable that the attempted logging will not be
successful either during times like these, but even when things are this
bad do not let the server to die; it might become responsive later.

Needless to say for external status requester this error will look like
system is down, and that makes the server not to be considered healty.

Reference: https://github.com/kerolasa/f5gs/issues/4
Reported-by: Markus Elfring <elfring@users.sourceforge.net>
Signed-off-by: Sami Kerola <kerolasa@iki.fi>