update status file immediately after opening it
The open will blank the file, and if there does not happen to be status
update the file remains blank resulting to a loss of status.  This is
pretty bad bug, and should result to a new release sometime soon.

Broken-since: 5d74d9bac9fe757217dc12f027eb9c6ba72a28ca
Signed-off-by: Sami Kerola <kerolasa@iki.fi>