do not let daemon to stop when system is running out of memory
When system is short of memory to the extent allocating enough space
for couple file descriptors and pointers is not possible it is unlikely
the subsequent logging will work.  Oh well, the important bit is that
daemon keeps on going no matter what, and of course for applications
checking the state this problem means system is effectively not usable;
that is actually is assuming the condition being true.

Signed-off-by: Sami Kerola <kerolasa@iki.fi>