add get_quiet_server_status() to verify state change did happen
For some reason on systems that has ran f5gs for long time the server
reacts slower to ipc message actions than socket requests, resulting the
post-state change output to differ with requested state.  To counter
weight make the post-change validation run for multiple times until it is
the requested state, or fail with an error message.

Signed-off-by: Sami Kerola <kerolasa@iki.fi>